Adopt two review suggestions: a back-reference and a tighter type
From Codacy's review of #73. flush()'s docstring now points at
teardown._flush_buffer the way that function already points back, so a
future change to the drain-then-append protocol finds both copies from
either end. And fold_interval's return type narrows to
tuple[float, int] | None — intervals is a record count in both branches,
and the wider float claim was an artifact of extraction, not a fact
about the value.

The review's third suggestion — a test for ascii_fallback under an
invalid encoding name — is already covered:
test_the_frames_degrade_to_what_the_encoding_carries parametrizes over
"definitely-not-a-codec", which is the LookupError path.

+++ src/lumberjack/renderers/progress/smoothing.py
@@ -32,7 +32,7 @@ def _smoothed(known: float | None, sample: float) -> float:
 
 def fold_interval(
     previous: float | None, newest: float, oldest: float, count: int
-) -> tuple[float, float] | None:
+) -> tuple[float, int] | None:
     """The `(span, intervals)` one delta of records contributes, or None.
 
     Two cases, both exact rather than approximate:
